requirements.txt
black==26.1.0
django-dump-load-utf8==0.0.4
django-phonenumber-field==8.4.0
phonenumbers==9.0.22
pipdeptree==2.30.0
psycopg2-binary==2.9.11
pylint-django==2.7.0
python-dotenv==1.2.1
